requirements.txt
requests>=2.31.0
PyNaCl>=1.5.0